refactor: extract SegmentedDownloadHelper and fix all broken references
- Extract shared download loop (downloadSegments/downloadBatch) into
  SegmentedDownloadHelper, eliminating ~200 lines of duplicated code
  between HttpDownloadSource and FtpDownloadSource
- Both sources now delegate to SegmentedDownloadHelper.downloadAll()
  with a protocol-specific segment download lambda
- Fix headers→properties rename across all KetchApi implementations:
  RemoteKetch, InstanceManager, FakeKetchApi, ResolveUrlRequest,
  ServerRoutes
- Update all test fake DownloadSource objects: rename headers→properties,
  add buildResumeState() implementation
- Update DefaultFileNameResolverTest to use ResolvedSource instead of
  ServerInfo
- Remove obsolete TaskRecordTest for removed HTTP-specific fields
- Update DownloadContext KDoc to be protocol-agnostic
