+++ library/endpoints/src/commonMain/kotlin/com/linroid/ketch/endpoints/model/ResolveUrlRequest.kt
@@ -6,10 +6,10 @@ import kotlinx.serialization.Serializable
  * Request body for resolving a URL without downloading.
  *
  * @property url the URL to resolve
- * @property headers optional HTTP headers to include in the probe
+ * @property properties source-specific key-value pairs (e.g., HTTP headers)
  */
 @Serializable
 data class ResolveUrlRequest(
   val url: String,
-  val headers: Map<String, String> = emptyMap(),
+  val properties: Map<String, String> = emptyMap(),
 )
